TheAI/src/automated_penetration_testing.py
```python
import os
from scapy.all import *
from pyNMAP import pyNMAP

class AutomatedPenetrationTesting:
    def __init__(self):
        self.pen_test_results = []

    def automate_pen_testing(self, target_ip):
        nm = pyNMAP()
        scan_results = nm.nmap_ping_scan(target_ip)

        for result in scan_results:
            if result['status'] == 'up':
                self.exploit_vulnerabilities(result['ip'])

        return self.pen_test_results

    def exploit_vulnerabilities(self, target_ip):
        # This is a placeholder function. In a real-world application, this function would contain code to exploit known vulnerabilities.
        # For ethical and legal reasons, this code is not included here.
        pass

    def save_results(self, filename):
        with open(filename, 'w') as file:
            for result in self.pen_test_results:
                file.write(str(result) + '\n')

if __name__ == "__main__":
    apt = AutomatedPenetrationTesting()
    apt.automate_pen_testing('192.168.1.0/24')
    apt.save_results('pen_test_results.txt')
```