TheAI/src/cyber_security_training.py
```python
import os
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score
from sklearn.preprocessing import LabelEncoder

class CyberSecurityTraining:
    def __init__(self, data_path):
        self.data_path = data_path
        self.employee_training_material = None

    def load_data(self):
        if os.path.exists(self.data_path):
            # Load the data from a CSV file
            data = pd.read_csv(self.data_path)
            return data
        else:
            raise FileNotFoundError(f"{self.data_path} does not exist.")

    def preprocess_data(self, data):
        # Preprocess the data
        label_encoder = LabelEncoder()
        for col in data.columns:
            if data[col].dtype == 'object':
                data[col] = label_encoder.fit_transform(data[col])
        return data

    def train_model(self):
        # Load and preprocess the data
        data = self.load_data()
        data = self.preprocess_data(data)

        # Split the data into training and testing sets
        X_train, X_test, y_train, y_test = train_test_split(data.drop('label', axis=1), data['label'], test_size=0.2, random_state=42)

        # Train the model
        model = RandomForestClassifier(n_estimators=100)
        model.fit(X_train, y_train)

        # Evaluate the model
        y_pred = model.predict(X_test)
        accuracy = accuracy_score(y_test, y_pred)
        print(f"Model accuracy: {accuracy}")

        # Save the model
        self.employee_training_material = model

    def train_employees(self):
        if self.employee_training_material is not None:
            print("Training employees...")
            # Training logic goes here
            print("Training completed.")
        else:
            print("No training material available.")

if __name__ == "__main__":
    training = CyberSecurityTraining("data/cyber_security_training_data.csv")
    training.train_model()
    training.train_employees()
```