TheAI/requirements.txt
tensorflow==2.4.1
pytorch==1.7.1
scikit-learn==0.24.1
nltk==3.5
spacy==3.0.5
nmap==0.6.1
scapy==2.4.4
pycrypto==2.6.1
pandas==1.2.3
numpy==1.20.1
matplotlib==3.3.4
flask==1.1.2
sqlalchemy==1.4.3
pymongo==3.11.3
pytest==6.2.2
unittest2==1.1.0
